streamlit/requirements.txt
boto3
strands-agents
strands-agents-tools
dotenv
requests
streamlit